[dev] Resume, Recruit Controller인가기능 추가 ( userId를 직접 입력하는 방식 -> 로그인된 유저 정보 이용 방식 )

+++ src/main/java/com/carero/controller/RecruitApiController.java
@@ -18,10 +18,15 @@ import io.swagger.annotations.ApiParam;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AuthorizationServiceException;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @RestController
@@ -39,7 +44,7 @@ public class RecruitApiController {
     public RecruitReadDto readRecruit(
             @PathVariable("id") Long id
     ){
-        Recruit recruit = recruitService.findOne(id);
+        Recruit recruit = recruitService.findById(id);
         return new RecruitReadDto(recruit);
     }
 
@@ -50,7 +55,7 @@ public class RecruitApiController {
             @ApiParam(value = "한 페이지 당 보여질 개수")@RequestParam(value = "limit", defaultValue = "8") int limit
     ){
         List<RecruitPageDto> recruits = recruitService.findByPage(page*8, limit);
-        int count = recruitService.countAll();
+        long count = recruitService.countAll();
 
         return new ResultPaging(count, page, recruits);
     }
@@ -63,12 +68,18 @@ public class RecruitApiController {
             @PathVariable("id") Long id
             ){
 
-        User user = userService.findById(recruitCUDRequestDto.getUserId());
+        User user = userService.getMyUser()
+                .orElseThrow(() -> new UsernameNotFoundException("현재 로그인된 정보를 가져오지 못했습니다."));
+        Recruit nowRecruit = recruitService.findById(id);
 
-        List<SubCategory> subCats = getSubCategories(recruitCUDRequestDto);
+        if (nowRecruit.getUser() == user) {
+            List<SubCategory> subCats = getSubCategories(recruitCUDRequestDto);
 
-        Recruit newRecruit = recruitCUDRequestDto.createRecruit(user,subCats);
-        recruitService.update(id, newRecruit);
+            Recruit newRecruit = recruitCUDRequestDto.createRecruit(user,subCats);
+            recruitService.update(id, newRecruit);
+        }else{
+            throw new AuthorizationServiceException("해당 글을 수정할 권한이 없습니다.");
+        }
 
         return new RecruitCUDResponseDto(id);
     }
@@ -78,7 +89,16 @@ public class RecruitApiController {
     @DeleteMapping("/{id}")
     public ResponseEntity<RecruitCUDResponseDto> deleteRecruit(
             @PathVariable("id") Long id){
-        recruitService.deleteById(id);
+
+        User user = userService.getMyUser()
+                .orElseThrow(() -> new UsernameNotFoundException("현재 로그인된 정보를 가져오지 못했습니다."));
+        Recruit target = recruitService.findById(id);
+
+        if (target.getUser() == user ){
+            recruitService.deleteById(id);
+        } else{
+            throw new AuthorizationServiceException("해당 글을 삭제할 권한이 없습니다.");
+        }
 
         return new ResponseEntity<>(new RecruitCUDResponseDto(id), HttpStatus.OK);
     }
@@ -87,7 +107,10 @@ public class RecruitApiController {
     @PreAuthorize("hasAnyRole('USER','ADMIN')")
     @PostMapping
     public ResponseEntity<RecruitCUDResponseDto> createRecruit(@RequestBody RecruitCUDRequestDto recruitCUDRequestDto){
-        User user = userService.findById(recruitCUDRequestDto.getUserId());
+
+        User user = userService.getMyUser()
+                .orElseThrow(() -> new UsernameNotFoundException("현재 로그인된 정보를 가져오지 못했습니다."));
+
         if(user == null){
             throw new IllegalStateException("해당 이름의 유저는 없습니다.");
         }
